requirements.txt
pytest
python-dotenv
openai
pandas
matplotlib
streamlit